Fit a Linear or Mixed Effects Model with Optional Modifiers

This function fits a linear model (`lm`) or a mixed effects model (`lmer`) depending on the presence of random effects. It supports interaction terms, sensitivity covariates, and returns a tidy summary of the model. Users can choose whether to compute p-values for mixed models. It also supports looping over multiple outcomes and exposures.

Arguments

data

Your dataset containing the variables used in the model.

outcome

A string or character vector specifying one or more outcome variables.

exposure

A string or character vector specifying one or more exposure variables.

covariates

A character vector of covariate names to adjust for.

effect_modifier

A string specifying an effect modifier to interact with the exposure.

sensitivity_cov

A character vector of sensitivity covariates.

random_effects

A string specifying the random effects structure (e.g., "(1 | group)").

p_values

Logical. If TRUE (default), uses lmerTest to compute p-values for mixed models. If FALSE, uses lme4 without p-values.

Value

A list containing model results. If multiple outcomes or exposures are provided, returns a named list of results.

Examples

# Examples

data <- mtcars

# 1. Run a single model with one outcome and one exposure
run_linear_models(data = data,
                  outcome = "mpg",
                  exposure = "cyl")
#> $model
#> 
#> Call:
#> (function (formula, data, subset, weights, na.action, method = "qr", 
#>     model = TRUE, x = FALSE, y = FALSE, qr = TRUE, singular.ok = TRUE, 
#>     contrasts = NULL, offset, ...) 
#> {
#>     ret.x <- x
#>     ret.y <- y
#>     cl <- match.call()
#>     mf <- match.call(expand.dots = FALSE)
#>     m <- match(c("formula", "data", "subset", "weights", "na.action", 
#>         "offset"), names(mf), 0L)
#>     mf <- mf[c(1L, m)]
#>     mf$drop.unused.levels <- TRUE
#>     mf[[1L]] <- quote(stats::model.frame)
#>     mf <- eval(mf, parent.frame())
#>     if (method == "model.frame") 
#>         return(mf)
#>     else if (method != "qr") 
#>         warning(gettextf("method = '%s' is not supported. Using 'qr'", 
#>             method), domain = NA)
#>     mt <- attr(mf, "terms")
#>     y <- model.response(mf, "numeric")
#>     w <- as.vector(model.weights(mf))
#>     if (!is.null(w) && !is.numeric(w)) 
#>         stop("'weights' must be a numeric vector")
#>     offset <- model.offset(mf)
#>     mlm <- is.matrix(y)
#>     ny <- if (mlm) 
#>         nrow(y)
#>     else length(y)
#>     if (!is.null(offset)) {
#>         if (!mlm) 
#>             offset <- as.vector(offset)
#>         if (NROW(offset) != ny) 
#>             stop(gettextf("number of offsets is %d, should equal %d (number of observations)", 
#>                 NROW(offset), ny), domain = NA)
#>     }
#>     if (is.empty.model(mt)) {
#>         x <- NULL
#>         z <- list(coefficients = if (mlm) matrix(NA_real_, 0, 
#>             ncol(y)) else numeric(), residuals = y, fitted.values = 0 * 
#>             y, weights = w, rank = 0L, df.residual = if (!is.null(w)) sum(w != 
#>             0) else ny)
#>         if (!is.null(offset)) {
#>             z$fitted.values <- offset
#>             z$residuals <- y - offset
#>         }
#>     }
#>     else {
#>         x <- model.matrix(mt, mf, contrasts)
#>         z <- if (is.null(w)) 
#>             lm.fit(x, y, offset = offset, singular.ok = singular.ok, 
#>                 ...)
#>         else lm.wfit(x, y, w, offset = offset, singular.ok = singular.ok, 
#>             ...)
#>     }
#>     class(z) <- c(if (mlm) "mlm", "lm")
#>     z$na.action <- attr(mf, "na.action")
#>     z$offset <- offset
#>     z$contrasts <- attr(x, "contrasts")
#>     z$xlevels <- .getXlevels(mt, mf)
#>     z$call <- cl
#>     z$terms <- mt
#>     if (model) 
#>         z$model <- mf
#>     if (ret.x) 
#>         z$x <- x
#>     if (ret.y) 
#>         z$y <- y
#>     if (!qr) 
#>         z$qr <- NULL
#>     z
#> })(formula = .Primitive("quote")(mpg ~ cyl), data = structure(list(
#>     mpg = c(21, 21, 22.8, 21.4, 18.7, 18.1, 14.3, 24.4, 22.8, 
#>     19.2, 17.8, 16.4, 17.3, 15.2, 10.4, 10.4, 14.7, 32.4, 30.4, 
#>     33.9, 21.5, 15.5, 15.2, 13.3, 19.2, 27.3, 26, 30.4, 15.8, 
#>     19.7, 15, 21.4), cyl = c(6, 6, 4, 6, 8, 6, 8, 4, 4, 6, 6, 
#>     8, 8, 8, 8, 8, 8, 4, 4, 4, 4, 8, 8, 8, 8, 4, 4, 4, 8, 6, 
#>     8, 4), disp = c(160, 160, 108, 258, 360, 225, 360, 146.7, 
#>     140.8, 167.6, 167.6, 275.8, 275.8, 275.8, 472, 460, 440, 
#>     78.7, 75.7, 71.1, 120.1, 318, 304, 350, 400, 79, 120.3, 95.1, 
#>     351, 145, 301, 121), hp = c(110, 110, 93, 110, 175, 105, 
#>     245, 62, 95, 123, 123, 180, 180, 180, 205, 215, 230, 66, 
#>     52, 65, 97, 150, 150, 245, 175, 66, 91, 113, 264, 175, 335, 
#>     109), drat = c(3.9, 3.9, 3.85, 3.08, 3.15, 2.76, 3.21, 3.69, 
#>     3.92, 3.92, 3.92, 3.07, 3.07, 3.07, 2.93, 3, 3.23, 4.08, 
#>     4.93, 4.22, 3.7, 2.76, 3.15, 3.73, 3.08, 4.08, 4.43, 3.77, 
#>     4.22, 3.62, 3.54, 4.11), wt = c(2.62, 2.875, 2.32, 3.215, 
#>     3.44, 3.46, 3.57, 3.19, 3.15, 3.44, 3.44, 4.07, 3.73, 3.78, 
#>     5.25, 5.424, 5.345, 2.2, 1.615, 1.835, 2.465, 3.52, 3.435, 
#>     3.84, 3.845, 1.935, 2.14, 1.513, 3.17, 2.77, 3.57, 2.78), 
#>     qsec = c(16.46, 17.02, 18.61, 19.44, 17.02, 20.22, 15.84, 
#>     20, 22.9, 18.3, 18.9, 17.4, 17.6, 18, 17.98, 17.82, 17.42, 
#>     19.47, 18.52, 19.9, 20.01, 16.87, 17.3, 15.41, 17.05, 18.9, 
#>     16.7, 16.9, 14.5, 15.5, 14.6, 18.6), vs = c(0, 0, 1, 1, 0, 
#>     1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 
#>     0, 1, 0, 1, 0, 0, 0, 1), am = c(1, 1, 1, 0, 0, 0, 0, 0, 0, 
#>     0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 
#>     1, 1, 1, 1), gear = c(4, 4, 4, 3, 3, 3, 3, 4, 4, 4, 4, 3, 
#>     3, 3, 3, 3, 3, 4, 4, 4, 3, 3, 3, 3, 3, 4, 5, 5, 5, 5, 5, 
#>     4), carb = c(4, 4, 1, 1, 2, 1, 4, 2, 2, 4, 4, 3, 3, 3, 4, 
#>     4, 4, 1, 2, 1, 1, 2, 2, 4, 2, 1, 2, 2, 4, 6, 8, 2)), row.names = c("Mazda RX4", 
#> "Mazda RX4 Wag", "Datsun 710", "Hornet 4 Drive", "Hornet Sportabout", 
#> "Valiant", "Duster 360", "Merc 240D", "Merc 230", "Merc 280", 
#> "Merc 280C", "Merc 450SE", "Merc 450SL", "Merc 450SLC", "Cadillac Fleetwood", 
#> "Lincoln Continental", "Chrysler Imperial", "Fiat 128", "Honda Civic", 
#> "Toyota Corolla", "Toyota Corona", "Dodge Challenger", "AMC Javelin", 
#> "Camaro Z28", "Pontiac Firebird", "Fiat X1-9", "Porsche 914-2", 
#> "Lotus Europa", "Ford Pantera L", "Ferrari Dino", "Maserati Bora", 
#> "Volvo 142E"), class = "data.frame"))
#> 
#> Coefficients:
#> (Intercept)          cyl  
#>      37.885       -2.876  
#> 
#> 
#> $tidy
#> # A tibble: 2 × 9
#>   term        estimate conf.low conf.high std.error  p.value error n_obs   BIC
#>   <chr>          <dbl>    <dbl>     <dbl>     <dbl>    <dbl> <lgl> <int> <dbl>
#> 1 (Intercept)    37.9     33.6      42.1      2.07  8.37e-18 NA       32  174.
#> 2 cyl            -2.88    -3.53     -2.22     0.322 6.11e-10 NA       32  174.
#> 
#> $formula
#> mpg ~ cyl
#> <environment: 0x56087128c038>
#> 
#> $residuals
#>  [1]  0.3701643  0.3701643 -3.5814159  0.7701643  3.8217446 -2.5298357
#>  [7] -0.5782554 -1.9814159 -3.5814159 -1.4298357 -2.8298357  1.5217446
#> [13]  2.4217446  0.3217446 -4.4782554 -4.4782554 -0.1782554  6.0185841
#> [19]  4.0185841  7.5185841 -4.8814159  0.6217446  0.3217446 -1.5782554
#> [25]  4.3217446  0.9185841 -0.3814159  4.0185841  0.9217446 -0.9298357
#> [31]  0.1217446 -4.9814159
#> 
#> $exposure
#> [1] "cyl"
#> 
#> attr(,"class")
#> [1] "run_model_result"

# 2. Store the output for a single model
model_output <- run_linear_models(data = data,
                                  outcome = "mpg",
                                  exposure = "cyl")
model_output$tidy  # access tidy results
#> # A tibble: 2 × 9
#>   term        estimate conf.low conf.high std.error  p.value error n_obs   BIC
#>   <chr>          <dbl>    <dbl>     <dbl>     <dbl>    <dbl> <lgl> <int> <dbl>
#> 1 (Intercept)    37.9     33.6      42.1      2.07  8.37e-18 NA       32  174.
#> 2 cyl            -2.88    -3.53     -2.22     0.322 6.11e-10 NA       32  174.

# 3. Store output for multiple outcomes and exposures
multi_output <- run_linear_models(data = data,
                                  outcome = c("mpg", "hp"),
                                  exposure = c("cyl", "wt"))
multi_output[["mpg&cyl"]]$tidy  # access a specific model's tidy results
#> NULL

# 4. Use named vectors for outcomes and exposures
outcomes <- c("mpg", "hp")
names(outcomes) <- outcomes
exposures <- c("cyl", "wt")
names(exposures) <- exposures
named_output <- run_linear_models(data = data,
                                  outcome = outcomes,
                                  exposure = exposures)
named_output[["hp&wt"]]$tidy
#> NULL

# 5. Include covariates directly
covar_output <- run_linear_models(data = data,
                                  outcome = "mpg",
                                  exposure = "cyl",
                                  covariates = c("wt", "hp"))
covar_output$tidy
#> # A tibble: 4 × 9
#>   term        estimate conf.low conf.high std.error  p.value error n_obs   BIC
#>   <chr>          <dbl>    <dbl>     <dbl>     <dbl>    <dbl> <lgl> <int> <dbl>
#> 1 (Intercept)  38.8     35.1     42.4        1.79   4.80e-19 NA       32  163.
#> 2 wt           -3.17    -4.68    -1.65       0.741  1.99e- 4 NA       32  163.
#> 3 hp           -0.0180  -0.0424   0.00629    0.0119 1.40e- 1 NA       32  163.
#> 4 cyl          -0.942   -2.07     0.187      0.551  9.85e- 2 NA       32  163.

# 6. Include covariates using a named vector
covars <- c("wt", "hp")
names(covars) <- covars
named_covar_output <- run_linear_models(data = data,
                                        outcome = "mpg",
                                        exposure = "cyl",
                                        covariates = covars)
named_covar_output$tidy
#> # A tibble: 4 × 9
#>   term        estimate conf.low conf.high std.error  p.value error n_obs   BIC
#>   <chr>          <dbl>    <dbl>     <dbl>     <dbl>    <dbl> <lgl> <int> <dbl>
#> 1 (Intercept)  38.8     35.1     42.4        1.79   4.80e-19 NA       32  163.
#> 2 wt           -3.17    -4.68    -1.65       0.741  1.99e- 4 NA       32  163.
#> 3 hp           -0.0180  -0.0424   0.00629    0.0119 1.40e- 1 NA       32  163.
#> 4 cyl          -0.942   -2.07     0.187      0.551  9.85e- 2 NA       32  163.

# 7. Include an effect modifier (e.g., interaction term)
interaction_output <- run_linear_models(data = data,
                                        outcome = "mpg",
                                        exposure = "cyl",
                                        effect_modifier = "am")
interaction_output$tidy
#> # A tibble: 4 × 10
#>   term        estimate conf.low conf.high std.error  p.value error n_obs   BIC
#>   <chr>          <dbl>    <dbl>     <dbl>     <dbl>    <dbl> <lgl> <int> <dbl>
#> 1 (Intercept)    30.9     24.3     37.4       3.19  1.95e-10 NA       32  173.
#> 2 cyl            -1.98    -2.89    -1.06      0.449 1.41e- 4 NA       32  173.
#> 3 am             10.2      1.36    19.0       4.30  2.53e- 2 NA       32  173.
#> 4 cyl:am         -1.31    -2.75     0.143     0.707 7.55e- 2 NA       32  173.
#> # ℹ 1 more variable: modifier <chr>